Improvement: Extended connection limit to 30. Internaly more connection search is performed
This commit sets the connection limit up to 30. This is done by additionaly search more
connection from a existing Connection reference.

Note: The limit of 30 is only a hard coded limit to avoid a DoS attack via the transport api

* Changed the upper limit from 6 to 30 connection
** Added possiblity to search more connection based on a existing connection
* Changed the SBB request URI

Important!!
The change was necessary due the fact, that the xmlfahrplan API is not capable to perfom more
connection requests